Change the interval for process statistic auto-logging to microsecond resolution, although expecting log entries at a rate of one per microsecond is in the realm of fantasy. However, several per second is reasonable, and changing the interval matches the units used for Watchdog timers as well. Change Watchdog to use the Time:: defined value for the interval instead of hardcoding microseconds/second. Update the test program so it still autologs at one-per-second.
[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 2472]

+++ src/include/be_process_statistics.h
@@ -173,15 +173,19 @@ namespace BiometricEvaluation {
 			/**
 			 * @brief
 			 * Start logging process statistics automatically,
-			 * in intervals of seconds. The first log entry will
-			 * occur soon after the call to this method as the
-			 * delay interval is invoked after the first entry.
+			 * in intervals of microseconds. The first log entry
+			 * will occur soon after the call to this method as
+			 * the delay interval is invoked after the first entry.
+			 * @note
+			 * It is unrealistic to expect that log entries can
+			 * be made at a rate of one per microsecond.
 			 * @note
 			 * If stopAutoLogging() is called very soon after the
 			 * start, a log entry may not be made.
 			 *
 			 * @param interval[in]
-			 *	The gap between logging snapshots, in seconds.
+			 *	The gap between logging snapshots, in
+			 *	microseconds.
 			 * @throw Error::ObjectDoesNotExist
 			 *	The LogSheet does not exist; this object was
 			 *	not created with LogCabinet object.
@@ -193,7 +197,7 @@ namespace BiometricEvaluation {
 			 *	The statistics gathering is not implemented for
 			 *	this operating system.
 			 */
-			void startAutoLogging(int interval)
+			void startAutoLogging(uint64_t interval)
 			    throw (Error::ObjectDoesNotExist,
 				Error::ObjectExists,
 				Error::StrategyError,
